Modularize http api endpoints for server
